Testing pattern figured out
Ok so the deal is: if you use wd.promiseChainRemote() you can return the
browser object after each mocha it() block and you don't have to call
notify(done) or tell mocha that it is async...it just does it! But watch
out because mocha will pass the test as true without the return statement
and you'll never know something was wrong.

* const changed to var. Since there's no ES6 elsewhere in the project, I
don't see any point in starting now.